Needed to move functions to be tested to init, not main, for tests to be caught.
